add authentication login

+++ backend/controllers/productController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
-const Product = require('../models/productModel')
+const Product = require('../models/productModel');
+const User = require('../models/userModel');
 // @desc Get products
 // @route GET /api/products
 // @access Private
 const getProducts = asyncHandler(async (req, res) => {
 
-    const products = await Product.find()
+    const products = await Product.find({user: req.user.id})
     res.status(200).json(products)
 })
 // @desc Set products
@@ -17,7 +18,8 @@ const setProducts = asyncHandler(async (req, res) => {
         throw new Error('Please enter a text field')
     }
     const product = await Product.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(product)
 })
@@ -29,6 +31,19 @@ const updateProducts = asyncHandler(async (req, res) => {
     if (!product) {
         res.status(400);
         throw new Error('Product not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the product user
+    if (product.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     const updateProduct = await Product.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -43,6 +58,19 @@ const deleteProducts = asyncHandler(async (req, res) => {
     if (!product) {
         res.status(400);
         throw new Error('Product not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the product user
+    if (product.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
         await Product.deleteOne()
     res.status(200).json({id: req.params.id})
